Adjustments made to the test location and robustness.
